Refactor out HandsProvider<T>
